requirements.txt
unsloth>=2026.2.0
unsloth_zoo>=2026.2.0
torch>=2.5.0
transformers>=5.0.0,<6.0.0
trl>=0.25.0
datasets>=3.3.0
accelerate>=1.5.0
peft>=0.17.0
bitsandbytes>=0.47.0
sentencepiece>=0.2.0
safetensors>=0.6.0
tokenizers>=0.22.0
numpy>=2.2.0
pandas>=2.3.0
pyyaml>=6.0.2
tqdm>=4.67.0
Pillow>=10.4.0
torchvision>=0.20.0
jsonschema>=4.23.0
langdetect>=1.0.9
nbformat>=5.10.4
ipykernel>=6.30.1
